# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Connecting Neighbors - Hyperlocal Community App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: integer
        replitId:
          type: string
        name:
          type: string
        username:
          type: string
        bio:
          type: string
          nullable: true
        apartment:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        isVerified:
          type: boolean
        joinedAt:
          type: string
          format: date-time
      required: [id, replitId, name, username, isVerified, joinedAt]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        apartment:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
...